+++ src/grammar-support/log-manager.js
@@ -1,76 +1,5 @@
+const BranchList = require('./branch-list');
 const TagList = require('./tag-list');
-
-class BranchList {
-  constructor() {
-    /**
-     * @type {Map<string, Branch>}
-     * @private
-     */
-    this._branches = new Map();
-  }
-
-  /**
-   * @param {string} branch
-   * @throws {Error}
-   */
-  add(branch) {
-    if (this._branches.has(branch)) {
-      throw new Error('Branch already exists: ' + branch);
-    }
-
-    this._branches.set(branch, new Branch(branch));
-  }
-
-  /**
-   * @param {string} branch
-   * @return {Branch}
-   * @throws {Error}
-   */
-  get(branch) {
-    if (!this._branches.has(branch)) {
-      throw new Error('Branch not created: ' + branch);
-    }
-
-    return this._branches.get(branch);
-  }
-}
-
-class Branch {
-  /**
-   * @param {string} name
-   */
-  constructor(name) {
-    /**
-     * @type {string}
-     * @private
-     */
-    this._name = name;
-
-    /**
-     * @type {number}
-     * @private
-     */
-    this._commits = 0;
-  }
-
-  /**
-   * @returns {string}
-   */
-  getName() {
-    return this._name;
-  }
-
-  incrementCommitCount() {
-    this._commits += 1;
-  }
-
-  /**
-   * @returns {number}
-   */
-  getCommitCount() {
-    return this._commits;
-  }
-}
 
 class LogManager {
   constructor() {
